SKY-14-redirect-rules-test Refactor: removing PageFactory and production pages
