+++ src/test/java/com/mentoring/LoginTest.java
@@ -2,12 +2,12 @@ package com.mentoring;
 
 import com.mentoring.framework.BasicTest;
 import com.mentoring.model.User;
+import com.mentoring.pageobject.MainPage;
+import com.mentoring.pageobject.NotificationContainer;
 import com.mentoring.util.Messages;
 import com.mentoring.util.CommonAssertions;
 import com.mentoring.framework.utils.UserUtils;
 import com.mentoring.pageobject.CharacterSelectionPage;
-import com.mentoring.pageobject.NotificationContainer;
-import com.mentoring.pageobject.PageFactory;
 import io.qameta.allure.Feature;
 import io.qameta.allure.Severity;
 import io.qameta.allure.SeverityLevel;
@@ -36,12 +36,12 @@ public class LoginTest extends BasicTest {
     @Feature(LOGIN)
     public void successfulLogin() {
         User user = new User();
-        CharacterSelectionPage characterSelectionPage = new PageFactory().getMainPage(driver)
+        CharacterSelectionPage characterSelectionPage = new MainPage(driver)
                 .login(user);
-
         softAssertion.assertThat(characterSelectionPage.isLogOutButtonVisible())
                 .as("Logout button should be visible.")
                 .isTrue();
+
         softAssertion.assertAll();
     }
 
@@ -49,13 +49,13 @@ public class LoginTest extends BasicTest {
     @Severity(SeverityLevel.BLOCKER)
     @Feature(LOGIN)
     public void wrongUserName() {
-        User user = new User();
-        new PageFactory().getMainPage(driver)
+        new MainPage(driver)
                 .fillLoginForm(UserUtils.generateRandomUsername(), VALID_PASSWORD)
                 .clickLoginButton();
 
-        NotificationContainer notificationContainer = new PageFactory().getNotificationContainer(driver);
+        NotificationContainer notificationContainer = new NotificationContainer(driver);
         CommonAssertions.assertNotificationMessageIsCorrect(softAssertion, notificationContainer, Messages.BAD_CREDENTIALS);
+
         softAssertion.assertAll();
     }
 
@@ -64,12 +64,13 @@ public class LoginTest extends BasicTest {
     @Feature(LOGIN)
     public void wrongPassword() {
         User user = new User();
-        new PageFactory().getMainPage(driver)
+        new MainPage(driver)
                 .fillLoginForm(user.getUserName(), WRONG_PASSWORD)
                 .clickLoginButton();
 
-        NotificationContainer notificationContainer = new PageFactory().getNotificationContainer(driver);
+        NotificationContainer notificationContainer = new NotificationContainer(driver);
         CommonAssertions.assertNotificationMessageIsCorrect(softAssertion, notificationContainer, Messages.BAD_CREDENTIALS);
+
         softAssertion.assertAll();
     }
 
@@ -78,12 +79,13 @@ public class LoginTest extends BasicTest {
     @Feature(LOGIN)
     public void emptyPassword() {
         User user = new User();
-        new PageFactory().getMainPage(driver)
+        new MainPage(driver)
                 .fillLoginForm(user.getUserName(), "")
                 .clickLoginButton();
 
-        NotificationContainer notificationContainer = new PageFactory().getNotificationContainer(driver);
+        NotificationContainer notificationContainer = new NotificationContainer(driver);
         CommonAssertions.assertNotificationMessageIsCorrect(softAssertion, notificationContainer, Messages.EMPTY_CREDENTIALS);
+
         softAssertion.assertAll();
     }
 
@@ -91,12 +93,13 @@ public class LoginTest extends BasicTest {
     @Severity(SeverityLevel.BLOCKER)
     @Feature(LOGIN)
     public void emptyUsername() {
-        new PageFactory().getMainPage(driver)
+        new MainPage(driver)
                 .fillLoginForm("", VALID_PASSWORD)
                 .clickLoginButton();
 
-        NotificationContainer notificationContainer = new PageFactory().getNotificationContainer(driver);
+        NotificationContainer notificationContainer = new NotificationContainer(driver);
         CommonAssertions.assertNotificationMessageIsCorrect(softAssertion, notificationContainer, Messages.EMPTY_CREDENTIALS);
+
         softAssertion.assertAll();
     }
 
@@ -104,12 +107,13 @@ public class LoginTest extends BasicTest {
     @Severity(SeverityLevel.NORMAL)
     @Feature(LOGIN)
     public void emptyFields() {
-        new PageFactory().getMainPage(driver)
+        new MainPage(driver)
                 .fillLoginForm("", "")
                 .clickLoginButton();
 
-        NotificationContainer notificationContainer = new PageFactory().getNotificationContainer(driver);
+        NotificationContainer notificationContainer = new NotificationContainer(driver);
         CommonAssertions.assertNotificationMessageIsCorrect(softAssertion, notificationContainer, Messages.EMPTY_CREDENTIALS);
+
         softAssertion.assertAll();
     }
 }
