--- src/test/java/com/mentoring/pageobject/PageFactory.java
@@ -1,19 +1,0 @@
-package com.mentoring.pageobject;
-
-import com.mentoring.framework.Config;
-import org.openqa.selenium.WebDriver;
-
-public class PageFactory {
-
-    public MainPage getMainPage(WebDriver driver) {
-        return Config.isLocalEnvironmentUsed() ? new MainPageDev(driver) : new MainPageProd(driver);
-    }
-
-    public CharacterSelectionPage getCharacterSelectionPage(WebDriver driver) {
-        return Config.isLocalEnvironmentUsed() ? new CharacterSelectionPageDev(driver) : new CharacterSelectionPageProd(driver);
-    }
-
-    public NotificationContainer getNotificationContainer(WebDriver driver) {
-        return Config.isLocalEnvironmentUsed() ? new NotificationContainerDev(driver) : new NotificationContainerProd(driver);
-    }
-}
